+++ src/editor.js
@@ -18,6 +18,7 @@ import {appliedState, batchedApplier} from './utils/applied-state';
 import {assert} from '../lib/assert';
 import {attachBlobUrl, FilesDragHint, fileSortCompare} from './file-upload';
 import {classMap} from 'lit-html/directives/class-map';
+import {CTA_TYPES} from './cta-types';
 import {Deferred} from '../vendor/ampproject/amphtml/src/utils/promise';
 import {getNamespace} from '../lib/namespace';
 import {hintIgnoreEnds, hintsUrl, setAttrFileHints} from './hints';
@@ -38,7 +39,7 @@ import {
   viewportIdFull,
   ViewportSelector,
 } from './viewport';
-import AmpStoryAdPreview, {CTA_TYPES} from './amp-story-ad-preview';
+import AmpStoryAdPreview from './amp-story-ad-preview';
 import codemirror from '../lib/runtime-deps/codemirror';
 import fs from 'fs-extra';
 import htmlMinifier from 'html-minifier';
@@ -647,7 +648,6 @@ class Editor {
     const doc = this.codeMirror_.getValue();
     const docWithFileRefs = this.replaceFileRefs_(doc);
     this.preview_.update(docWithFileRefs);
-    this.preview_.setMetaCtaLabel_(docWithFileRefs);
   }
 
   toggleFullPreview_() {
@@ -718,6 +718,10 @@ class Editor {
       delete htmlSchema[key];
     }
     Object.assign(htmlSchema, hints[format.toLowerCase()]);
+
+    // Below, ours:
+    // - Sets attribute values for `meta[content]` to hint CTA
+    // - Sets uploaded file hints
     htmlSchema.meta.attrs.content = Object.keys(CTA_TYPES);
     this.updateFileHints_();
   }
